FIX issue #20. For ME sums float is not enough, we need double!
time ./gcheck.exe -p 16384 32 120
***************************************
NumIterations             = 120
NumThreadsPerBlock        = 32
NumBlocksPerGrid          = 16384
---------------------------------------
FP precision              = DOUBLE (nan=0)
Complex type              = THRUST::COMPLEX
Momenta memory layout     = AOSOA[32]
Wavefunction GPU memory   = LOCAL
Curand generation         = DEVICE (CUDA code)
---------------------------------------
NumberOfEntries           = 120
TotalTimeInWaveFuncs      = 9.715026e-02 sec
MeanTimeInWaveFuncs       = 8.095855e-04 sec
StdDevTimeInWaveFuncs     = 7.475417e-06 sec
MinTimeInWaveFuncs        = 8.022210e-04 sec
MaxTimeInWaveFuncs        = 8.058030e-04 sec
---------------------------------------
NumMatrixElementsComputed = 62914560
MatrixElementsPerSec      = 6.476006e+08 sec^-1
***************************************
NumMatrixElements(notNan) = 62914560
MeanMatrixElemValue       = 1.371798e-02 GeV^0
StdErrMatrixElemValue     = 1.033853e-06 GeV^0
StdDevMatrixElemValue     = 8.200386e-03 GeV^0
MinMatrixElemValue        = 6.071582e-03 GeV^0
MaxMatrixElemValue        = 3.374925e-02 GeV^0
***************************************
00 CudaFree : 0.121930 sec
0a ProcInit : 0.000514 sec
0b MemAlloc : 0.181631 sec
0c GenCreat : 0.014437 sec
1a GenSeed  : 0.000097 sec
1b GenRnGen : 0.077821 sec
2a RamboIni : 0.000893 sec
2b RamboFin : 0.000505 sec
2c CpDTHwgt : 0.065954 sec
2d CpDTHmom : 0.703514 sec
3a SigmaKin : 0.000799 sec
3b CpDTHmes : 0.096351 sec
4a DumpLoop : 0.216250 sec
9a DumpAll  : 0.234272 sec
9b GenDestr : 0.000239 sec
9c MemFree  : 0.017983 sec
9d CudReset : 0.039325 sec
TOTAL       : 1.772515 sec
TOTAL(n-2)  : 1.611260 sec
***************************************
real    0m1.783s
user    0m1.133s
sys     0m0.648s

+++ examples/gpu/eemumu_AV/SubProcesses/P1_Sigma_sm_epem_mupmum/check.cc
@@ -193,7 +193,7 @@ int main(int argc, char **argv)
   hstMEs = new fptype[nMEs]();
 #endif
 
-  float* wavetimes = new float[niter]();
+  double* wavetimes = new double[niter]();
   fptype* matrixelementvector = new fptype[niter * ndim * process.nprocesses]();
 
 #ifdef __CUDACC__
@@ -301,7 +301,7 @@ int main(int argc, char **argv)
     // 3b. Copy MEs back from device to host
 
     // *** START THE OLD TIMER ***
-    float gputime = 0;
+    double gputime = 0;
 
     // --- 3a. SigmaKin
     const std::string skinKey = "3a SigmaKin";
@@ -430,10 +430,10 @@ int main(int argc, char **argv)
 
   if (perf)
   {
-    float sum = 0;
-    float sq_sum = 0;
-    float mintime = wavetimes[0];
-    float maxtime = wavetimes[0];
+    double sum = 0;
+    double sq_sum = 0;
+    double mintime = wavetimes[0];
+    double maxtime = wavetimes[0];
     for (int iiter = 0; iiter < niter; ++iiter)
     {
       sum += wavetimes[iiter];
@@ -441,15 +441,15 @@ int main(int argc, char **argv)
       mintime = std::min( mintime, wavetimes[iiter] );
       maxtime = std::max( mintime, wavetimes[iiter] );
     }
-    float mean = sum / niter;
-    float stdev = std::sqrt( sq_sum / niter - mean * mean );
+    double mean = sum / niter;
+    double stdev = std::sqrt( sq_sum / niter - mean * mean );
 
     const int num_mes = niter*ndim;
     int num_nan = 0;
-    float sumelem = 0;
-    float sqselem = 0;
-    float minelem = matrixelementvector[0];
-    float maxelem = matrixelementvector[0];
+    double sumelem = 0;
+    double sqselem = 0;
+    double minelem = matrixelementvector[0];
+    double maxelem = matrixelementvector[0];
     for (int imes = 0; imes < num_mes; ++imes)
     {
       if ( isnan( matrixelementvector[imes] ) )
@@ -461,11 +461,11 @@ int main(int argc, char **argv)
       }
       sumelem += matrixelementvector[imes];
       sqselem += matrixelementvector[imes]*matrixelementvector[imes];
-      minelem = std::min( minelem, (float)matrixelementvector[imes] );
-      maxelem = std::max( maxelem, (float)matrixelementvector[imes] );
+      minelem = std::min( minelem, (double)matrixelementvector[imes] );
+      maxelem = std::max( maxelem, (double)matrixelementvector[imes] );
     }
-    float meanelem = sumelem / ( num_mes - num_nan );
-    float stdelem = std::sqrt( sqselem / ( num_mes - num_nan) - meanelem * meanelem );
+    double meanelem = sumelem / ( num_mes - num_nan );
+    double stdelem = std::sqrt( sqselem / ( num_mes - num_nan) - meanelem * meanelem );
 
     std::cout << "***************************************" << std::endl
               << "NumIterations             = " << niter << std::endl
